Working on memory / throughput optimization