@Aaronontheweb 2021-05-28
Actor MemoryFootprint values from `dev`:

``` ini

BenchmarkDotNet=v0.13.0, OS=Windows 10.0.19041.985 (2004/May2020Update/20H1)
AMD Ryzen 7 1700, 1 CPU, 16 logical and 8 physical cores
.NET SDK=5.0.203
  [Host]     : .NET Core 3.1.15 (CoreCLR 4.700.21.21202, CoreFX 4.700.21.21402), X64 RyuJIT
  Job-LTLHWP : .NET Core 3.1.15 (CoreCLR 4.700.21.21202, CoreFX 4.700.21.21402), X64 RyuJIT

IterationCount=25  RunStrategy=Monitoring  WarmupCount=5  

```
|     Method | SpawnCount |     Mean |    Error |   StdDev |      Gen 0 |     Gen 1 | Gen 2 | Allocated |
|----------- |----------- |---------:|---------:|---------:|-----------:|----------:|------:|----------:|
| SpawnActor |      10000 | 238.0 ms | 22.09 ms | 29.49 ms | 12000.0000 | 3000.0000 |     - |     70 MB |