@Aaronontheweb 2021-06-03
I don't really want to include the latest code in Akka.NET v1.4.21 because it is.... Extremely ugly and probably needs to be refactored, but as far as I can tell it works and is significantly faster than its predecessor

``` ini

BenchmarkDotNet=v0.13.0, OS=Windows 10.0.19041.985 (2004/May2020Update/20H1)
AMD Ryzen 7 1700, 1 CPU, 16 logical and 8 physical cores
.NET SDK=5.0.203
  [Host]     : .NET Core 3.1.15 (CoreCLR 4.700.21.21202, CoreFX 4.700.21.21402), X64 RyuJIT
  DefaultJob : .NET Core 3.1.15 (CoreCLR 4.700.21.21202, CoreFX 4.700.21.21402), X64 RyuJIT


```
|                                     Method |      Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------------------------------------- |----------:|---------:|---------:|-------:|------:|------:|----------:|
|                            ActorPath_Parse | 516.62 ns | 6.548 ns | 6.125 ns | 0.1068 |     - |     - |     448 B |
|                           ActorPath_Concat |  58.44 ns | 1.244 ns | 2.809 ns | 0.0267 |     - |     - |     112 B |
|                           ActorPath_Equals |  20.04 ns | 0.135 ns | 0.105 ns |      - |     - |     - |         - |
|                         ActorPath_ToString | 183.06 ns | 3.714 ns | 8.682 ns | 0.0267 |     - |     - |     112 B |
|            ActorPath_ToSerializationFormat | 192.53 ns | 2.282 ns | 1.906 ns | 0.0267 |     - |     - |     112 B |
| ActorPath_ToSerializationFormatWithAddress | 196.93 ns | 3.905 ns | 5.845 ns | 0.0267 |     - |     - |     112 B |