@Aaronontheweb 2021-06-03
`ActorPath_Parse` has gotten faster with each of these changes, and I am pretty sure I know why. Will add those changes to a separate PR.