@Aaronontheweb 2021-06-03
Test suite indicates there's some problems with the parsing formula. Going to let this sit for now.