@Aaronontheweb 2021-09-02
Latest numbers, after fixing how UID encoding works:

``` ini

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.19041.1165 (2004/May2020Update/20H1)
AMD Ryzen 7 1700, 1 CPU, 16 logical and 8 physical cores
.NET SDK=5.0.302
  [Host]     : .NET Core 3.1.17 (CoreCLR 4.700.21.31506, CoreFX 4.700.21.31502), X64 RyuJIT
  DefaultJob : .NET Core 3.1.17 (CoreCLR 4.700.21.31506, CoreFX 4.700.21.31502), X64 RyuJIT


```
|                                     Method |      Mean |    Error |   StdDev |  Gen 0 | Allocated |
|------------------------------------------- |----------:|---------:|---------:|-------:|----------:|
|                            ActorPath_Parse | 483.99 ns | 7.935 ns | 7.422 ns | 0.1068 |     448 B |
|                           ActorPath_Concat |  48.10 ns | 0.095 ns | 0.084 ns | 0.0268 |     112 B |
|                           ActorPath_Equals |  18.87 ns | 0.058 ns | 0.054 ns |      - |         - |
|                         ActorPath_ToString | 175.05 ns | 2.630 ns | 2.460 ns | 0.0267 |     112 B |
|            ActorPath_ToSerializationFormat | 185.83 ns | 1.485 ns | 1.389 ns | 0.0267 |     112 B |
| ActorPath_ToSerializationFormatWithAddress | 194.60 ns | 3.157 ns | 2.953 ns | 0.0267 |     112 B |