@Aaronontheweb 2021-09-02
Looking at the `RemotePingPong` benchmarks - still looks like there's some cases not handled by the stringifier here